requirements.txt
asgiref==3.7.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.2.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
googlemaps==4.10.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==0.21.1
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.2.3
